fix(restoreState): Handle girder post-processing
Similar to recent changes in fileLoader regarding loading of labelmaps, we need
to convert vtkImageData that has been annotated (via girder) with the labelmap
designation.
